fix type

+++ tasks/mpi/korovin_n_line_topology/func_tests/main.cpp
@@ -51,7 +51,7 @@ TEST(korovin_n_line_topology_mpi, transfer_data) {
     for (int i = 0; i < n; i++) {
       ASSERT_EQ(received_data[i], data[i]);
     }
-    for (int i = 0; i < received_trajectory.size(); i++) {
+    for (int i = 0; i < (int)received_trajectory.size(); i++) {
       ASSERT_EQ(received_trajectory[i], root + i);
     }
   }
@@ -103,7 +103,7 @@ TEST(korovin_n_line_topology_mpi, transfer_data_random) {
     for (int i = 0; i < n; i++) {
       ASSERT_EQ(received_data[i], data[i]);
     }
-    for (int i = 0; i < received_trajectory.size(); i++) {
+    for (int i = 0; i < (int)received_trajectory.size(); i++) {
       ASSERT_EQ(received_trajectory[i], root + i);
     }
   }
